requirements.txt
flask>=3.0
pillow>=10.0
numpy>=1.26
reportlab>=4.0
bcrypt>=4.1
PyJWT>=2.8
python-dotenv>=1.0
gunicorn>=21.0
mollie-api-python>=3.0
